PUB-2002 - Updated README, seperated out tests
